fix: update dependencies and versions in package files
Update the version of `electron-to-chromium` to `1.5.54` and 
add `@tanstack/table-core` as a new dependency with version 
`^8.20.5`. Upgrade `vaul-svelte` to `1.0.0-next.2` and adjust 
its peer dependencies. Modify the version of `cross-spawn` to 
`7.0.5` and update the version of `paneforge` to `1.0.0-next.1`. 
These changes ensure compatibility with the latest features and 
improvements in the respective packages.